added mantine, refactored loadingSpinner
