+++ index.html
@@ -4,12 +4,9 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <link rel="icon" type="image/svg+xml" href="/facivon.png" />
-    <meta
-      property="og:image"
-      itemprop="image"
-      content="/images/public.jpg"
-    />
+    <meta property="og:image" itemprop="image" content="/images/public.jpg" />
     <link href="https://fonts.googleapis.com/css2?family=Heebo&display=swap" rel="stylesheet" />
+    <script src="https://cdn.jsdelivr.net/npm/chart.js"></script>
 
     <title>חידון חכמים</title>
   </head>
