+++ src/routes/Layout.tsx
@@ -1,14 +1,15 @@
 import { Outlet } from "react-router-dom";
 import Navbar from "./Navbar";
-import Loader from "~/Componenets/Loader";
 import "./App.scss";
+import useLoadingSpinner from "~/hooks/useLoadingSpinner";
 
 export default function Layout() {
+  const { LoadingSpinnerModal } = useLoadingSpinner();
   return (
     <>
       <Navbar />
-      <Loader />
       <Outlet />
+      <LoadingSpinnerModal />
       <footer>© עמותת שבושון</footer>
     </>
   );
